+++ app/widgets/index.js
@@ -15,3 +15,4 @@ export * from './TableSelector';
 export * from './TextEntry';
 export * from './TimeRange';
 export * from './VisualStimulusResponse';
+export * from './RadioPrizes';