+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.14.51] - 2021-01-05
 - Fixed notification time
+- Added TokenLogger Prizes Activity feature
 
 ## [0.14.50] - 2021-01-05
 - Removed console logs